// scripts/totals/interfaces/runner-skip.ts
/**
 * @license
 * Copyright (c) Noldova.
 *
 * This source code is licensed under the license found in the
 * LICENSE file in the root directory of this source tree.
 */

export default interface IRunnerSkip {
  readonly file: string;
  readonly names: readonly string[];
  readonly reason: string;
}